,parallel.sh: add warning

+++ bin/,parallel.sh
@@ -33,7 +33,7 @@ megs=$(( size / 1024 / 1024 ))
 len=$(( megs / procs ))
 
 if [ $len -lt 1 ]; then
-    echo >&2 "less than 1 M per job"
+    echo >&2 "WARNING: less than 1M per job, not parallelizing"
     "$@" < "$file"
     exit $?
 fi
@@ -51,25 +51,29 @@ for i in ${!ps[@]}; do
     start[$i]=$offset
     [ $i -gt 0 ] || continue
     offset=$(( offset * 1048576 ))
-    adjustment[$i]=$(dd if="$file" bs=1 skip=$offset \
+    adjustment[$i]=$(dd if="$file" bs=1 skip=$offset count=2048 2>/dev/null \
       | head -n1 | wc -c | awk '{ print $1 }')
-    byteoffs[$i]=$(( $offset + 0 ))
+    [ ${adjustment[$i]} -lt 2048 ] || {
+        echo >&2 "WARNING: newlines sparse, may affect line-oriented commands"
+        adjustment[$i]=0
+    }
+    byteoffs[$i]=$offset
 done
 adjustment[0]=0
-byteoffs[0]=0
 
 job() (
     i=$1; shift 1
 
-    [ $i -lt $((procs-1)) ] && count="count=$((len))" || count=""
+    [ $i -lt $((procs-1)) ] && count="count=$len" || count=""
 
     (
         dd if="$file" bs=1024k skip=${start[$i]} $count 2>/dev/null \
         | (
-            dd bs=1 of=/dev/null count=${adjustment[$i]} 2>/dev/null
+            [ ${adjustment[$i]} -eq 0 ] || \
+                dd bs=1 of=/dev/null count=${adjustment[$i]} 2>/dev/null
             cat
         )
-        if [ $i -lt $((procs-1)) ]; then
+        if [ $i -lt $((procs-1)) ] && [ ${adjustment[$((i+1))]} -gt 0 ]; then
             dd if="$file" bs=1 skip=${byteoffs[$((i+1))]} \
               count=${adjustment[$((i+1))]} 2>/dev/null
         fi
@@ -90,4 +94,14 @@ wait
 # option to abort all jobs when one fail?
 # option to abort all jobs when one finishes?
 # option to return exit status based on all/some/none of the jobs?
+#
+# -p num  set number of procs
+#
+# -e      end on first error (fail fast)
+# -m num  max completions, -m1 : "any"
+#
+# exit status:
+# default: 0 if all jobs succeed (like xargs -P)
+# -s  some success - 0 if at least one job succeeds
+#
